+++ src/backend/InvenTree/order/models.py
@@ -1685,6 +1685,7 @@ class OrderLineItem(InvenTree.models.InvenTreeMetadataModel):
     Attributes:
         quantity: Number of items
         reference: Reference text (e.g. customer reference) for this line item
+        project_code: Project code associated with this line item (optional)
         note: Annotation for the item
         target_date: An (optional) date for expected shipment of this line item.
     """
@@ -1766,6 +1767,15 @@ class OrderLineItem(InvenTree.models.InvenTreeMetadataModel):
         help_text=_(
             'Target date for this line item (leave blank to use the target date from the order)'
         ),
+    )
+
+    project_code = models.ForeignKey(
+        common_models.ProjectCode,
+        on_delete=models.SET_NULL,
+        blank=True,
+        null=True,
+        verbose_name=_('Project Code'),
+        help_text=_('Select project code for this order'),
     )
 
 
@@ -1992,8 +2002,6 @@ class PurchaseOrderExtraLine(OrderExtraLine):
 
     Attributes:
         order: Link to the PurchaseOrder that this line belongs to
-        title: title of line
-        price: The unit price for this OrderLine
     """
 
     class Meta:
@@ -2422,8 +2430,6 @@ class SalesOrderExtraLine(OrderExtraLine):
 
     Attributes:
         order: Link to the SalesOrder that this line belongs to
-        title: title of line
-        price: The unit price for this OrderLine
     """
 
     class Meta: